+++ packages/zengine-ui-react/src/molecules/TextField.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactElement } from 'react';
 import { useField } from 'formik';
 
 import Input from '../atoms/Input';
@@ -23,8 +23,8 @@ export interface TextFieldProps {
   innerRef?: any
   label?: string
   labelClasses?: string
-  prefix?: string
-  suffix?: string
+  prefix?: string | ReactElement
+  suffix?: string | ReactElement
   type?: string
 }
 
@@ -92,7 +92,7 @@ const TextField: React.FC<TextFieldProps> = (props) => {
 
           { input }
 
-          { !isEmpty(props.prefix) ? (
+          { !isEmpty(props.suffix) ? (
             <div className="input-group-prepend"><span className="input-group-text">{ props.suffix }</span></div>
           ) : undefined }
         </div>