+++ packages/zengine-ui-react/src/util/Form.tsx
@@ -4,6 +4,7 @@ import isEmpty from 'lodash/isEmpty';
 import isObject from 'lodash/isObject';
 
 import Button from '../atoms/Button';
+import FocusFormErrors from './FocusFormErrors';
 
 interface FormProps {
   enableReinitialize?: boolean
@@ -30,7 +31,7 @@ interface FormikProperties {
   dirty: boolean
   isValid: boolean
   isSubmitting: boolean
-  errors: object
+  errors: { [key: string]: string }
   touched: { [key: string]: boolean }
   values: object
 }
@@ -96,13 +97,20 @@ function ZengineUIForm (props: FormProps): React.ReactElement {
       validateOnChange={validateOnChange}
       onSubmit={async (values, actions) => {
         actions.setSubmitting(true);
-        const res = await onSubmit?.(values);
-        // React to submit callback return value.
-        if (isObject(res)) {
-          actions.setErrors(res);
-        } else {
+        try {
+          await onSubmit?.(values);
           actions.resetForm();
           afterSubmit?.(values);
+        } catch (err) {
+          if (isObject(err)) {
+            if (err instanceof Error) {
+              actions.setErrors({ _form: err.message })
+            } else {
+              actions.setErrors(err);
+            }
+          } else {
+            actions.setErrors({ _form: err })
+          }
         }
 
         actions.setSubmitting(false);
@@ -116,8 +124,12 @@ function ZengineUIForm (props: FormProps): React.ReactElement {
             {props.children}
 
             {/* If the form has errors, display a message above buttons. */}
-            {!isEmpty(errors) && allFieldsTouched(touched) ? (
+            {!isEmpty(errors) && (allFieldsTouched(touched) || !isEmpty(errors._form)) ? (
               <div className="invalid-feedback d-block mb-2">
+                {!isEmpty(errors._form) ? (
+                  <><span>{errors._form}</span><br /></>
+                ) : null}
+
                 Please fix the above errors and try again.
               </div>
             ) : null}
@@ -165,6 +177,7 @@ function ZengineUIForm (props: FormProps): React.ReactElement {
                 {isSubmitting && <p className="mb-0 text-info">{saveMessage}</p>}
               </div>
             )}
+            <FocusFormErrors />
           </Form>
         )
       }}