+++ packages/zengine-ui-react/stories/Welcome.stories.js
@@ -145,8 +145,8 @@ function MyComponent() {
 export const Development = () => (
   <article>
     <PageTitle>Developing Components</PageTitle>
-    <p>Be sure to run <code>npm run scss</code> at least once so you have the generated stylesheet for the Storybook
-      preview</p>
+    <p>Be sure to use the non-minified stylesheet so you have the generated stylesheet for the Storybook
+      preview. Optionally, run <code>npm run watch</code> in the <code>zengine-ui</code> package to get live reload for theming.</p>
 
     <p>There's 3 mandatory parts for every component: code, stories and tests.</p>
 
@@ -170,7 +170,7 @@ export const Development = () => (
       <ul>
         <li>A first "Default" story with minimum or no parameters</li>
         <li>If applicable, a "Style Guide" story displaying different configurations side by side</li>
-        <li>If applicable, a last "Playground" story using _Knobs_ to customize as many props as possible</li>
+        <li>If applicable, a last "Playground" story using <em>Knobs</em> to customize as many props as possible</li>
         <li>As many different stories as necessary to cover demonstrating key configurations of a component</li>
       </ul>
     </ul>