+++ packages/zengine-ui-react/stories/molecules/CountryField.stories.js
@@ -13,5 +13,11 @@ export default {
 };
 
 export const Default = () => (
-  <MockForm><CountryField label="CountryField" name="country" /></MockForm>
+  <MockForm>
+    <p>
+      This is a <em>SelectField</em> with custom, unchangeable options. All other SelectField props work here.
+      <br/>Data is pulled from the Zengine API.
+    </p>
+    <CountryField label="CountryField" name="country" />
+  </MockForm>
 );