+++ packages/zengine-ui-react/test/util/validation.test.ts
@@ -1,4 +1,18 @@
-import { exists, isEmpty, isString, matchesRegex, isEmail } from '../../src/util/validation';
+import {
+  exists,
+  isEmpty,
+  isString,
+  matchesRegex,
+  isEmail,
+  maxLength,
+  minLength,
+  maxWordCount,
+  minWordCount,
+  alpha,
+  alphaNumeric,
+  numeric,
+  zipCode,
+} from '../../src/util/validation';
 
 test('exists() correctly determines whether a value exists', () => {
   expect(exists(null)).toBe(false);
@@ -6,6 +20,8 @@ test('exists() correctly determines whether a value exists', () => {
   expect(exists(false)).toBe(true);
   expect(exists(0)).toBe(true);
   expect(exists('')).toBe(true);
+  expect(exists({})).toBe(true);
+  expect(exists([])).toBe(true);
 });
 
 test('isEmpty() correctly determines whether an array is empty', () => {
@@ -15,6 +31,7 @@ test('isEmpty() correctly determines whether an array is empty', () => {
 
 test('isEmpty() correctly determines whether an object is empty', () => {
   expect(isEmpty({})).toBe(true);
+  expect(isEmpty({ a: undefined })).toBe(false);
   expect(isEmpty({ a: 1 })).toBe(false);
 });
 
@@ -57,6 +74,69 @@ test('isEmail() correctly determines whether a value is an email address', () =>
   expect(isEmail('not an email')).toBe(false);
   expect(isEmail('without@domain')).toBe(false);
   expect(isEmail('withoutat.com')).toBe(false);
-  expect(isEmail('test@trivial.net')).toBe(true);
+  expect(isEmail('test@contrived.net')).toBe(true);
   expect(isEmail('test.foo.bar@trivial.net')).toBe(true);
+});
+
+test('maxLength() correctly determines whether a string adheres to a maximum length', () => {
+  expect(maxLength(123 as any, 123)).toBe(false);
+  expect(maxLength('hello', 3)).toBe(false);
+  expect(maxLength('hello', 5)).toBe(true);
+});
+
+test('minLength() correctly determines whether a string adheres to a maximum length', () => {
+  expect(minLength(123 as any, 123)).toBe(false);
+  expect(minLength('oi', 3)).toBe(false);
+  expect(minLength('oi', 2)).toBe(true);
+  expect(minLength('hello', 5)).toBe(true);
+});
+
+test('maxWordCount() correctly determines whether a string adheres to a maximum length', () => {
+  expect(maxWordCount('hello there foo bar', 3)).toBe(false);
+  expect(maxWordCount('hello there foo', 3)).toBe(true);
+});
+
+test('minWordCount() correctly determines whether a string adheres to a maximum length', () => {
+  expect(minWordCount('hello there', 3)).toBe(false);
+  expect(minWordCount('hello there foo', 3)).toBe(true);
+});
+
+test('alpha() correctly determines whether a string only contains alphabetic characters', () => {
+  expect(alpha('hello 123 there')).toBe(false);
+  expect(alpha('hellothere')).toBe(true);
+  expect(alpha('hello123')).toBe(false);
+  expect(alpha('HELLOthere')).toBe(true);
+  expect(alpha('HELLO there')).toBe(false);
+});
+
+test('alpha() optionally allows spaces', () => {
+  expect(alpha('hello 123', true)).toBe(false);
+  expect(alpha('hello there', true)).toBe(true);
+});
+
+test('alphaNumeric() correctly determines whether a string only contains alphabetic and numeric characters', () => {
+  expect(alphaNumeric('hello123there')).toBe(true);
+  expect(alphaNumeric('HELLOthere1')).toBe(true);
+  expect(alphaNumeric('hellOthere')).toBe(true);
+  expect(alphaNumeric('hellothere23{}')).toBe(false);
+  expect(alphaNumeric('hellothere!')).toBe(false);
+});
+
+test('alphaNumeric() optionally allows spaces', () => {
+  expect(alphaNumeric('hello 123 there', true)).toBe(true);
+});
+
+test('numeric() correctly determines whether a string only contains numeric characters', () => {
+  expect(numeric('hello 123 there')).toBe(false);
+  expect(numeric('123')).toBe(true);
+  expect(numeric('123 456')).toBe(false);
+  expect(numeric('123.456')).toBe(true);
+  expect(numeric('123!')).toBe(false);
+});
+
+test('zipCode() correctly determines whether a string is a valid US zip code', () => {
+  expect(zipCode('hello 123 there')).toBe(false);
+  expect(zipCode('9021')).toBe(false);
+  expect(zipCode('90210')).toBe(true);
+  expect(zipCode('902101')).toBe(false);
 });